Did some foundation GUI work with Spring, and gave it some testing. Seems to work nicely.
Added a method that produces an empty JFrame. This will allow me to create GUI windows with one method call.

Added a method that produces an empty JButton. This will allow me to create Buttons with one single method call.

Added a method that allows produces an empty JTextField. This will allow me to create uneditable text fields with ease.

Also documented each of these methods using javadocs documentation.

Now that the base work for the Homeroom GUI has gone in, the first piece of frontend to be developed is a log-in page, which will be used to access the DB.