Added a config file and can begin properly starting on a login system.
Wrote a very basic, and very hacky way of having a configuration file, which school administrators will be able to make use of in order to set the connection string for their MongoDB cluster. The implementations of methods in this class are not up to standard, and should be refactored appropriately when given time.

Need to do some basic testing to the login system once the code behind the database class has been written. Right now, it's about getting what can work, working, and refactoring and cleaning of the code can come later. A lot of perfective maintenance will need to be done throughout the entire program as it goes along.

Also begun planning multiple methods of connecting to a database, but for now, Homeroom will likely just be able to run off of a MongoDB cluster.