Did some basic refactoring work while (hopefully!) doing some more OO-thinking.
Refactored the GUIUtils.java class accordingly to ensure that a Window is now opened upon creation of the object. Methods can then be used on the object to fill and manipulate said Window with a little more ease. Also added a method to ensure that the Window object can also be directly accessed as and when I need to. Although, I'm generally hoping that I don't have to make use of that method unless absolutely necessary.