Made major additions to the program, along with slight minor fixes to other pieces of the program.
Added an "icon" picture to the jar file within Homeroom, allowing it to pull that picture from there to have as the main icon for the program.

Added the major backend and frontend for Student Management and student searching, and also made a very minor start on actual student data management.

Need to remember to make commits more often. Bad practise to just merge a singular branch with one commit inside.