Made some very basic code fixes, along with a few additions to some methods within some classes.
Also did some refactoring of code to allow for some minor QOL improvements. Also refactored the DBUtils class to ensure that connections are not uneccesarily made. The fixes made to this regard are not as good as they should be, but as far as minor improvements go, making a new connection every time a Student object was instantiated was not very good.