Made an extremely major start on Form Management.
The primary logic for Form Management has all been written and seems to work fine. What now needs to be done is debugging of a major issue, testing and documenting.

Experienced a major issue where the entire program would freeze up, waiting for a particular method call as the thread is blocking all processes until getting said input. While this may be intended behaviour, the JButtons (where aforementioned input is to be recieved from) does not seem to be properly appearing, meaning that it is not possible for input to be provided by the user. This then means that the entire program freezes up upon waiting for this.