Still experiencing the aforementioned major issue. Granted, there has not been much in the way of attempting to solve this issue, but there will be now.
I have opted to do some final additions to documentation that I left behind before finalising this branch, and making a second branch within this branch to work on the issue there.

Since the feature is purely supplementary to the Form Management class itself, but is a major addition to the rest of the program, I have decided to go to a new branch and work on everything I need to there. Once done, I will merge my changes with form branch, and continue working on this branch until it is time to delete forms branch by merging to main.

From there, I am hoping that for the sake of my deadline, I am able to finish this project and develop with no issues.