Made a minor start on the concurrency refactor.
Begun redesigning each and every class, and am now attempting to implement the new design into the rest of Homeroom. Provided these reworks to the code and the design work as intended, I can merge these with the forms branch and move on.

It is worth noting that I am only testing to ensure that the selection methods behind adding Students to Forms work. Anything else that may need to come after or before that can go.