Made a good start on implementing Teachers and their integration between Form Groups.
Managed to make a start on the bare minimum the Teacher class needs to start working with Forms Groups. All that really needs to be done now is the refactoring of the logic behind Forms to have them all working.

Of course, the database work is still more or less untested, which needs to be done.

This is a good start to be quite honest, should be able to finish the Teachers feature by Thursday, and after that I can move onto Classes.