Finished implementing Teachers and their integration between Form Groups.
Although the general code quality is not really up to standard, it is safe to say that the integration between Teachers and the rest of the Homeroom system is functional, and can be use to create form groups of students, with a teacher in charge.

Basic improvements to the entire program have been made, but more advanced improvements, which might require an entire branch worth of refactoring have been planned. As to how they will be implemented, I am not sure as to how and when.

As of right now, I do plan on continuing work with this project long after the deadline. So I will have to see how things go.

Will be merging this branch with the main branch, and will move onto creating the classes-branch, and all the work that will be going into that.