Finished the first iteration of 'Homeroom'.
Managed to implement the addition of all student in a Form to a Class.

Managed to implement the removal of a Student from Class.

Manged to implement the removal of a Student from a Form.

Managed to implement the removal of a Teacher from a Class.

Managed to implement the removal of a Teacher from a Form.